Moved to functions from trace_utilities.py to signal_processing.py. Add wrapper which return original functions + warning. Changed it everywhere. add doc-string
