+++ NuRadioReco/utilities/trace_utilities.py
@@ -1,4 +1,4 @@
-from NuRadioReco.utilities import units, ice, geometryUtilities as geo_utl, fft
+from NuRadioReco.utilities import units, fft, signal_processing
 import NuRadioReco.framework.base_trace
 
 import numpy as np
@@ -12,91 +12,48 @@ conversion_factor_integrated_signal = scipy.constants.c * scipy.constants.epsilo
 # see Phys. Rev. D DOI: 10.1103/PhysRevD.93.122005
 # to convert V**2/m**2 * s -> J/m**2 -> eV/m**2
 
-
-def get_efield_antenna_factor(station, frequencies, channels, detector, zenith, azimuth, antenna_pattern_provider):
-    """
-    Returns the antenna response to a radio signal coming from a specific direction
-
-    Parameters
-    ----------
-
-    station: Station
-    frequencies: array of complex
-        frequencies of the radio signal for which the antenna response is needed
-    channels: array of int
-        IDs of the channels
-    detector: Detector
-    zenith, azimuth: float, float
-        incoming direction of the signal. Note that refraction and reflection at the ice/air boundary are taken into account
-    antenna_pattern_provider: AntennaPatternProvider
-    """
-
-    efield_antenna_factor = np.zeros((len(channels), 2, len(frequencies)), dtype=complex)  # from antenna model in e_theta, e_phi
-    for iCh, channel_id in enumerate(channels):
-        zenith_antenna, t_theta, t_phi = geo_utl.fresnel_factors_and_signal_zenith(detector, station, channel_id, zenith)
-
-        if zenith_antenna is None:
-            logger.warning("Fresnel reflection at air-firn boundary leads to unphysical results, no reconstruction possible")
-            return None
-
-        logger.debug("angles: zenith {0:.0f}, zenith antenna {1:.0f}, azimuth {2:.0f}".format(
-            np.rad2deg(zenith), np.rad2deg(zenith_antenna), np.rad2deg(azimuth)))
-        antenna_model = detector.get_antenna_model(station.get_id(), channel_id, zenith_antenna)
-        antenna_pattern = antenna_pattern_provider.load_antenna_pattern(antenna_model)
-        ori = detector.get_antenna_orientation(station.get_id(), channel_id)
-        VEL = antenna_pattern.get_antenna_response_vectorized(frequencies, zenith_antenna, azimuth, *ori)
-        efield_antenna_factor[iCh] = np.array([VEL['theta'] * t_theta, VEL['phi'] * t_phi])
-
-    return efield_antenna_factor
+def get_efield_antenna_factor(*args, **kwargs):
+    logger.warning("get_efield_antenna_factor is moved to NuRadioReco.utilities.signal_processing.get_efield_antenna_factor")
+    return signal_processing.get_efield_antenna_factor(*args, **kwargs)
 
 
-def get_channel_voltage_from_efield(station, electric_field, channels, detector, zenith, azimuth, antenna_pattern_provider, return_spectrum=True):
-    """
-    Returns the voltage traces that would result in the channels from the station's E-field.
-
-    Parameters
-    ----------
-
-    station: Station
-    electric_field: ElectricField
-    channels: array of int
-        IDs of the channels for which the expected voltages should be calculated
-    detector: Detector
-    zenith, azimuth: float
-        incoming direction of the signal. Note that reflection and refraction
-        at the air/ice boundary are already being taken into account.
-    antenna_pattern_provider: AntennaPatternProvider
-    return_spectrum: boolean
-        if True, returns the spectrum, if False return the time trace
-    """
-
-    frequencies = electric_field.get_frequencies()
-    spectrum = electric_field.get_frequency_spectrum()
-    efield_antenna_factor = get_efield_antenna_factor(station, frequencies, channels, detector, zenith, azimuth, antenna_pattern_provider)
-    if return_spectrum:
-        voltage_spectrum = np.zeros((len(channels), len(frequencies)), dtype=complex)
-        for i_ch, ch in enumerate(channels):
-            voltage_spectrum[i_ch] = np.sum(efield_antenna_factor[i_ch] * np.array([spectrum[1], spectrum[2]]), axis=0)
-        return voltage_spectrum
-    else:
-        voltage_trace = np.zeros((len(channels), 2 * (len(frequencies) - 1)), dtype=complex)
-        for i_ch, ch in enumerate(channels):
-            voltage_trace[i_ch] = fft.freq2time(np.sum(efield_antenna_factor[i_ch] * np.array([spectrum[1], spectrum[2]]), axis=0), electric_field.get_sampling_rate())
-        return np.real(voltage_trace)
+def get_channel_voltage_from_efield(*args, **kwargs):
+    logger.warning("get_channel_voltage_from_efield is moved to NuRadioReco.utilities.signal_processing.get_channel_voltage_from_efield")
+    return signal_processing.get_channel_voltage_from_efield(*args, **kwargs)
 
 
 def get_electric_field_energy_fluence(electric_field_trace, times, signal_window_mask=None, noise_window_mask=None):
+    """ Compute the energy fluence of an electric field trace.
 
+    Parameters
+    ----------
+    electric_field_trace : 2d array
+        The electric field trace. The first dimension corresponds to the number of traces/polarizations,
+        the second dimension to the number of samples.
+    times : 1d array
+        The times corresponding to the electric field trace.
+    signal_window_mask : 1d array, default=None
+        A boolean mask to select the signal window.
+    noise_window_mask : 1d array, default=None
+        A boolean mask to select the noise window.
+
+    Returns
+    -------
+    energy_fluence : 1d array
+        The energy fluence of the electric field per polarization.
+    """
     if signal_window_mask is None:
         f_signal = np.sum(electric_field_trace ** 2, axis=1)
     else:
         f_signal = np.sum(electric_field_trace[:, signal_window_mask] ** 2, axis=1)
-    dt = times[1] - times[0]
+
     if noise_window_mask is not None and np.sum(noise_window_mask) > 0:
         f_noise = np.sum(electric_field_trace[:, noise_window_mask] ** 2, axis=1)
         f_signal -= f_noise * np.sum(signal_window_mask) / np.sum(noise_window_mask)
 
+    dt = times[1] - times[0]
     return f_signal * dt * conversion_factor_integrated_signal
+
 
 def get_stokes(trace_u, trace_v, window_samples=128, squeeze=True):
     """
@@ -174,6 +131,7 @@ def get_stokes(trace_u, trace_v, window_samples=128, squeeze=True):
         return np.squeeze(stokes)
     return stokes
 
+
 def upsampling_fir(trace, original_sampling_frequency, int_factor=2, ntaps=2 ** 7):
     """
     This function performs an upsampling by inserting a number of zeroes
@@ -249,7 +207,7 @@ def butterworth_filter_trace(trace, sampling_frequency, passband, order=8):
     n_samples = len(trace)
 
     spectrum = fft.time2freq(trace, sampling_frequency)
-    frequencies = np.fft.rfftfreq(n_samples, 1 / sampling_frequency)
+    frequencies = fft.freqs(n_samples, sampling_frequency)
 
     filtered_spectrum = apply_butterworth(spectrum, frequencies, passband, order)
     filtered_trace = fft.freq2time(filtered_spectrum, sampling_frequency)
@@ -340,7 +298,7 @@ def delay_trace(trace, sampling_frequency, time_delay, crop_trace=True):
     else:
         n_samples = len(trace)
         spectrum = fft.time2freq(trace, sampling_frequency)
-        frequencies = np.fft.rfftfreq(n_samples, 1 / sampling_frequency)
+        frequencies = fft.freqs(n_samples, sampling_frequency)
 
     spectrum *= np.exp(-1j * 2 * np.pi * frequencies * time_delay)
